+++ plugins/oauth/components/AuthenticatedUserAccountDropdown.js
@@ -2,44 +2,21 @@ import React from "react";
 
 import AuthStore from "#SRC/js/stores/AuthStore";
 import UserAccountDropdown from "#SRC/js/components/UserAccountDropdown";
-import UserAccountDropdownTrigger from "#SRC/js/components/UserAccountDropdownTrigger";
-import UserAccountDropdownTriggerContent from "#SRC/js/components/UserAccountDropdownTriggerContent";
 
 class AuthenticatedUserAccountDropdown extends UserAccountDropdown {
-  getUserLabel() {
-    const user = AuthStore.getUser();
-
-    let userLabel = null;
-
-    if (user && !user.is_remote) {
-      userLabel = user.description;
-    } else if (user && user.is_remote) {
-      userLabel = user.uid;
-    }
-
-    return userLabel;
-  }
-
   /**
    * Adds the authenticated user's account information to the dropdown items.
    * @return {array} Dropdown menu items.
    * @override
    */
   getMenuItems() {
-    const { clusterName, menuItems } = this.props;
+    const { menuItems, children } = this.props;
     const nextMenuItems = menuItems.slice();
     const userLabel = this.getUserLabel();
 
     nextMenuItems.unshift({
       className: "hidden",
-      html: (
-        <UserAccountDropdownTrigger>
-          <UserAccountDropdownTriggerContent
-            primaryContent={clusterName}
-            secondaryContent={userLabel}
-          />
-        </UserAccountDropdownTrigger>
-      ),
+      html: children,
       id: "dropdown-trigger",
       selectable: false
     });
